Updated ItemTypes
