+++ game/src/ecs/items/ItemType.java
@@ -1,9 +1,9 @@
 package ecs.items;
 
 public enum ItemType {
-    SwordFightingBasics,
-    Bread,
-    ExperiencePotion,
-    Bag,
+    BOOK,
+    FOOD,
+    POTION,
+    BAG,
     Basic
 }
