+++ game/src/ecs/items/explicitItems/Bread.java
@@ -1,26 +1,25 @@
-package ecs.items.ItemEntities;
+package ecs.items.explicitItems;
 
 import dslToGame.AnimationBuilder;
 import ecs.entities.Entity;
 import ecs.items.*;
 import tools.Point;
 
+public class Bread extends ItemData implements IOnCollect, IOnDrop, IOnUse {
 
-
-public class SwordFightingBasics extends ItemData implements IOnCollect, IOnDrop, IOnUse {
-
-    public SwordFightingBasics(){
+    public Bread(){
         super(
-            ItemType.SwordFightingBasics,
+            ItemType.FOOD,
             AnimationBuilder.buildAnimation("missingTexture.png"),
             AnimationBuilder.buildAnimation("missingTexture.png"),
-            "SwordFightingBasics",
-            "Adds a SwordSkill skill to the hero's SkillComponent");
+            "Bread",
+            "Heals the hero by 5 health points");
         this.setOnCollect(this);
         this.setOnDrop(this);
         this.setOnUse(this);
 
     }
+
     @Override
     public void onCollect(Entity WorldItemEntity, Entity whoCollides) {
